Changed route structure, Fixed session client update when user does auth actions.
